requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncstdlib-fw==3.13.2
attrs==25.3.0
betterproto-fw==2.0.3
bitsandbytes==0.46.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
fireworks-ai==0.17.16
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.73.1
grpclib==0.4.8
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.1
httpx-ws==0.7.2
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.2
jsonlines==4.0.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.2.1
nlpstats==0.0.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.74.0
packaging==24.2
pandas==2.2.3
peft==0.16.0
permpy==0.2.12
pillow==11.2.1
pluggy==1.5.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
replicate==1.0.7
requests==2.32.3
rich==14.0.0
ruff==0.9.10
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
trl==0.19.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
